Backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
audioop-lts==0.2.1
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
fastapi==0.116.1
h11==0.16.0
idna==3.10
numpy==2.3.2
pillow==11.3.0
pycryptodome==3.23.0
pydantic==2.11.7
pydantic_core==2.33.2
pydub==0.25.1
python-multipart==0.0.20
requests==2.32.4
scipy==1.16.1
sniffio==1.3.1
starlette==0.47.2
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
